# openapi.yaml
openapi: 3.0.3
info:
  title: Appliance Shop API - OpenAPI 3.0
  description: |-
    This is a Home Appliances Store Server based on the OpenAPI 3.0 specification. API for device management in Home Appliance Stores.

    Some useful links:
    - [The Appliance Shop repository](https://github.com/minhcopilot/Appliance-Shop-API)
    - [The source API definition for the Appliance Shop](https://github.com/minhcopilot/Appliance-Shop-API/openopenapi.yaml)

  contact:
    email: phanminh190502@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:9000
    description: Local server
...
paths:
...
  /customers:
    get:
      summary: Get all customers
      tags:
        - customer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '204':
          description: No content
        '500':
          description: Internal server error
    post:
      summary: Create a new customer
      tags:
        - customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCustomer'
      responses:
        '200':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Account already exists
        '500':
          description: Internal server error
  /customers/{id}:
    get:
      summary: Get customer by ID
      tags:
        - customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '410':
          description: Customer not found
        '500':
          description: Internal server error

    patch:
      summary: Update customer by ID
      tags:
        - customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
                address:
                  type: string
                birthday:
                  type: string
                  format: date
                email:
                  type: string
                password:
                  type: string
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Updated customer data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '410':
          description: Customer not found
        '500':
          description: Internal server error

    delete:
      summary: Delete customer by ID
      tags:
        - customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer deleted successfully
        '410':
          description: Customer not found
        '500':
          description: Internal server error
  /customers/change-password/{id}:
    patch:
      summary: Change customer password
      tags:
        - customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Invalid old password
        '404':
          description: User not found
        '500':
          description: Internal server error

  #Employee
...
  /orders/customer/{customerId}:
    get:
      tags:
        - order
      summary: Get orders by customer ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          schema:
            type: integer
          required: true
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        404:
          description: Customer not found
        500:
          description: Internal server error
...
  /vouchers/customer/{customerId}:
    get:
      tags:
        - voucher
      summary: Get vouchers by customer ID
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of vouchers for the customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Voucher'
        '404':
          description: Customer not found
        '500':
          description: Internal server error
    post:
      tags:
        - voucher
      summary: Create a voucher for a customer
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoucherInput'
      responses:
        '201':
          description: Voucher created for the customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
        '400':
          description: Bad request
        '404':
          description: Customer not found
        '500':
          description: Internal server error
  #Cart
...
  /chat/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - chat
      summary: Admin get all chat
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
              example:
                - id: '1'
                  employeeId: '101'
                - id: '2'
                  employeeId: null
        '500':
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Database Error'
  /chat/unassigned:
    get:
      security:
        - bearerAuth: []
      tags:
        - chat
      summary: Employee get unassigned chat
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
              example:
                - id: '2'
                  employeeId: null
        '500':
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Database Error'
  /chat/assigned:
    get:
      security:
        - bearerAuth: []
      tags:
        - chat
      summary: Employee get assigned chat
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
              example:
                - id: '1'
                  employeeId: '101'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Database Error'
  /chat/content/{id}:
    get:
      tags:
        - chat
      summary: Employee get messages by chat ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Chat ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
              example:
                - id: 'm1'
                  chatId: '1'
                  content: 'Hello'
                  timestamp: '2024-06-13T12:00:00Z'
                - id: 'm2'
                  chatId: '1'
                  content: 'How can I help you?'
                  timestamp: '2024-06-13T12:05:00Z'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Chat not found'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Database Error'
  /chat/message/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - chat
      summary: Employee get message by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Message ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: 'm1'
                chatId: '1'
                content: 'Hello'
                timestamp: '2024-06-13T12:00:00Z'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Message not found'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: 'Database Error'

  #Authentication User
...
components:
  schemas:
...
    NewCustomer:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        birthday:
          type: string
          format: date
        email:
          type: string
        password:
          type: string
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phoneNumber:
          type: string
          example: '123456789'
        email:
          type: string
          example: johndoe@example.com
...
    OrderDetail:
      type: object
      properties:
        productId:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        price:
          type: number
          example: 100.0
        discount:
          type: number
          example: 10
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        shippedDate:
          type: string
          format: date-time
          example: '2023-06-10T15:00:00Z'
        status:
          type: string
          example: 'PENDING'
        description:
          type: string
          example: 'Order description'
        shippingAddress:
          type: string
          example: '123 Street, City'
        shippingCity:
          type: string
          example: 'City'
        paymentType:
          type: string
          example: 'CREDIT_CARD'
        customerId:
          type: integer
          example: 1
        employeeId:
          type: integer
          example: 2
        orderDetails:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetail'
    Voucher:
      type: object
      properties:
        id:
          type: integer
          example: 1
        voucherCode:
          type: string
          example: 'DISCOUNT20'
        discountPercentage:
          type: number
          example: 20
        startDate:
          type: string
          format: date
          example: '2023-01-01'
        expiryDate:
          type: string
          format: date
          example: '2023-12-31'
        maxUsageCount:
          type: integer
          example: 100
        remainingUsageCount:
          type: integer
          example: 100
    VoucherInput:
      type: object
      properties:
        voucherCode:
          type: string
          example: 'DISCOUNT20'
        discountPercentage:
          type: number
          example: 20
        startDate:
          type: string
          format: date
          example: '2023-01-01'
        expiryDate:
          type: string
          format: date
          example: '2023-12-31'
        maxUsageCount:
          type: integer
          example: 100
      required:
        - voucherCode
        - discountPercentage
        - startDate
        - expiryDate
        - maxUsageCount
...
    Chat:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
          nullable: true
    Message:
      type: object
      properties:
        id:
          type: string
        chatId:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT